In article video poster image
Poster image is now woven into the video. This is at the mercy of whatever picture comes out of the API as the 'still image' field. We might want to have a chat with the people who upload them.

If the video is the first thing inside the body (i.e. the very first HTML element inside the body) then the main picture will be hidden.