Football fixtures upgraded to 'local' time.
We had a bug whereby we were using timezones incorrectly.

This has now been fixed, and dev machines will now run in a far away timezone so it is more obvious when you get a timezone wrong.

There is still some cleaning up and a bit of thinking around an abstraction to do, but since this is a World Cup fix i'm going to ship it first.