Fix the last edited heading in collections
Last modified time shows as 'x days ago ago'.
The commit also introcudes a way of running unit(?) tests for knockout.
The spec loads a mocked list of fronts, collections and items and starts the application.
The test is trivial, it just asserts the text of an element, but it can be extended in the future to run assertion on user interaction.