Update notification client
It provides the necessary changes in order to migrate the notification
service to Firebase. The Breaking news are now sent in a loop though
it is transparent as the interface remain the same.

Topics are now a list in order to keep the order. This has no impact on
breaking news but makes sense for other type of notifications (content)