use thumbnail for...thumbnail
Previously, we were using the master which is can be huge! The API response is providing a 140px wide image which doesn't look great in the 178px wide element, however I'll address this in another PR.

This should make the Tool that much faster as the client isn't downloading a 7MB+ image where there is an override in place.