Don't hydrate collection api from Grid
As I understand it, the `GET collection` will hydrate the `collection.json` with Grid image URLs from a URL in the json blob. I think a better solution to doing this would be to ensure the URL in the `collection.json` blob is correct at time of writing as this would stop us calling Grid each time we GET a collection. Write once, read many.