make v2 the default by redirecting
Will redirect if:

`redirect` query string is not present or is `true`

AND

(we're in DEV OR it's after Christmas 2019)

If `redirect` query string is not present, we default to `true`, however as it's not after Christmas 2019, we will not redirect. That is, there should be no impact on non DEV stages.