new table to track publication status of an issue version
When publishing, insert a row into `issue_versions` and `issue_versions_events`. This will allow us to record events fired by the Archiver lambda, providing the ability to see the progress of publication.

We can then render this on the client to provide confidence that the publishing action is doing something.