save Archiver messages into database
If we encounter a foreign key constraint violation, because the `version_id` doesn't exist in the `issue_versions` table, proceed to remove the message from SQS (that is, mark it as successful). This is because we can never recover from this state and it is the only way to progress.

We're only likely to enter this state in DEV and CODE as we share the same Archiver stack. That is, the SQS queue will have messages for DEV and CODE.