Dummy up an annual Edition
This is a playpen for Katy Vans and Chris Moran to see what the shape of a annual product might me. I've used special containers and fronts so that they can rename the sections as they see fit in the tool. There are no prefills or promotion metrics used, but if we did I've defaulted them to look at the preceding year. It's configured as a daily product, for testing, but clearly it isn't one.

I've not used the boiler plate special definitions, since we wanted 6 containers/collections in each front not the standard 3.

The product is likely to be mocked up with Film or Music content, so we've defaulted everything to the Culture swatch.

What did I get wrong??!??!