Extra fronts, with swatches set.
This looks mad, but as the special edition is 'lightly' defined, but we need to start production now, we're putting in more fronts than we need to cater for the colour scheme and the fact minds may change. 

We're anticipating having to change the order of fronts they use. There is a three colour scheme (neutral, lifestyle, opinion) and we are expecting 9 fronts, but potentially less. The additional fronts mean that we can reorder and rename in the tool, and be confident we can have the fronts meet the colour scheme.

We won't do it like this next time.