Update icon url
This is an updated icon for the forthcoming special edition. The icon has been passed through the Fastly image resizer that can be found (on dc1 vpn only) here: http://image-url-signer.s3-website-eu-west-1.amazonaws.com/ - the width is identical to the last icon, but it is 3px taller which we're not expecting to be a problem.