First stab at the next special edition
We don't know the swatches yet, the colours for the headers and text, the icon etc. This is a start for 10, this will be further modified before launch, once we have some more details. NB the huge amount of fronts is to cover us for the fact that we won't know until late in the day, what the ordering of swatches applied to fronts will.