Coronavirus front for AU
As requested by Merren:

"could we designate Front special 3 as the coronavirus container and have it prefill with coronavirus-tagged content? The top container could be coronavirus+Australia news, then the second container coronavirus+world/UK/US news, then the third can be coronavirus+opinion/comment. I'm not sure if we want to then strip corona-tagged content out of world, national and opinion, but that could be the next step."