Test should either return a promise or take a done callback
An asynchronous before/it/after function took a done callback but also returned a promise. This is not supported and will stop working in the future. Either remove the done callback (recommended) or change the function to not return a promise. (in spec: Fetch visible stories gets the visible stories)
Note: This message will be shown only once. Set the verboseDeprecations config property to true to see every occurrence.