docs: Use relative link for fronts-client
This docs link didn’t work for me locally: my editor interprets
`/fronts-client` as an absolute path and doesn’t take me to the right
file. Github does the right thing in the web UI for both styles, so this
commit changes it to use a path relative to the current directory.