Steal isError from tslint
Now that I’ve removed TSLint, this import is failing the build. To fix
the build, this commit copies the definition of isError from my local
node_modules.

I took a look at https://github.com/guardian/facia-tool/pull/1277, which
introduced this, and I didn’t see any explanation for this bit of code:
it seems strange to import something from tslint (which is a dev
dependency) for this, which makes me wonder if it was deliberate?

Either way, I believe the check is necessary: I initially tried removing
it, but that broke some tests.