Add formatting commit to .git-blame-ignore-revs
In a recent commit I ran prettier to format the code, which produced a
very big diff. This diff can get in the way of blaming with git, so Iʼve
recorded the commit ref in .git-blame-ignore-revs, which facilitates
telling git (using the config option blame.ignoreRevsFile) to ignore the
commit when blaming. (Iʼll add this config to the setup script shortly.)