Run scalafmt in CI
Now that this repo is using prettier to check formatting in CI, why not
also check scala formatting? This commit adds a check in CI that runs
scalafmt via an sbt plugin, copying the approach from
https://github.com/guardian/datawrapper-import/pull/12

I’ll also add a pre-commit hook shortly, and format the existing code.
I’ll probably choose a formatting style that uses tabs for indentation,
to be consistent with the frontend code (and given the accessibility
motivations for using tabs for indentation).