Update pre-commit hook to check scala formatting
This commit updates the pre-commit hook to run scalafmt the same way it
already runs prettier. Apart from the exit code tweak, this basically
involved copying the pre-commit hook from
https://github.com/guardian/datawrapper-import/pull/12

As with the prettier part of the hook, this should only run the
formatter on files with staged changes that don’t have unstaged changes,
and for any with both that fail formatting it should just report an
error and exit.

(I’ve also replaced the word “clobbering” with “overwriting”, because I
think it’s a less obscure phrasing.)