Prevent scalafmt check from blocking the build
Running scalafmt in CI is specifically about preventing _merges to main_
that contain formatting problems. There’s no need for it to prevent a
Riff Raff deployment being created if it fails, or block tests from
running.

Currently, if the code in a pull request works correctly but has a
formatting error, you must fix the formatting error in order to have
something you can deploy to CODE. This commit changes that, so that you
can deploy to CODE with formatting errors, but not merge to main.

(Hopefully this will provide a better developer experience!)

I’d like to avoid using sbt to run scalafmt, as sbt startup can be slow
and I’m worried it’ll require compiling the whole app, but I couldn’t
quickly find a way to install scalafmt on the runners that I was
confident of.