Get working build with sbt-assembly
For some reason, running `clean` in the build phase is essential to this
working! Without it it hangs, presumably somehow because sbt doesn’t
need to do anything.

I originally switched to sbt-assembly because the old packaging method
was hanging, but that may work now that I’ve realised I need `clean` in
there: will try that next.

I had thought having a default.nix would mean I could just use that for
the nix-shell, but it’s super slow. I think I’ll be sticking with the
shell.nix as well.

I should also expand to cover the startup scripts and similar next.