Just provide yarn with specific node version
This approach gives a nix shell with a yarn available which has the
right (major) node version for fronts-client (20), and a different
version available with node and npm (for the v1 app).

Unfortunately, nix refuses to allow node 16 as it’s EOL, so I’ve just
stuck with 18 for the moment in order to demonstrate that differences
are possible. Presumably there’s some way to get around the EOL warning,
but I’m not sure how right now.

With this much, it should be possible to write a script that runs each
of the commands in separate tmux terminals, for example. (Except for the
docker-compose bit, which I haven’t looked at yet.)