Provide command to run everything in tmux
This is a nice way to run all the commands and their output. My ideal
would be that we end up at an interactive shell where the command we
just ran is the most recent item in the history, and I can’t figure out
how to do this with tmux. Enabling remain-on-exit at least allows using
the respawn-pane command to restart the process if necessary.

I think it’d be nice (for myself) to have a way to run all of them in
emacs terminals, and probably a flexible way for others to run the
commands however they want.