Update code owner to Fronts and Curation
The [Fronts and Curation
team](https://github.com/orgs/guardian/teams/fronts-and-curation) are
now responsible for this repository, and I think they should be listed
as the code owner for it. This commit makes it so.

(digital-cms is a parent of that team which includes a lot of people,
who I think don’t necessarily need to get emails about reviewing PRs on
this repository.)